backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=5.0.1
celery>=5.3.6
boto3>=1.34.0
svix
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.27.0
python-multipart>=0.0.9
runpod
PyJWT>=2.8.0
